backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
python-decouple==3.8
Pillow==10.4.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
django-storages==1.14.2
